Extract view code in separate content view
