+++ GitHubSearch/Modules/Views/RepositoryInfoView.swift
@@ -1,0 +1,49 @@
+//
+//  RepositoryInfoView.swift
+//  GitHubSearch
+//
+//  Created by Eugene Karambirov on 01/02/2019.
+//  Copyright © 2019 Eugene Karambirov. All rights reserved.
+//
+
+import UIKit
+import SnapKit
+
+final class RepositoryInfoView: UIView {
+
+    lazy var nameLabel: UILabel = {
+        let label = UILabel()
+        label.font = UIFont.boldSystemFont(ofSize: 17)
+        label.numberOfLines = 0
+        return label
+    }()
+
+    lazy var descriptionLabel: UILabel = {
+        let label = UILabel()
+        label.font = UIFont.systemFont(ofSize: 15)
+        label.textColor = UIColor.gray
+        label.numberOfLines = 0
+        return label
+    }()
+
+    lazy var stackView: UIStackView = {
+        let stackView = UIStackView(arrangedSubviews: [nameLabel, descriptionLabel])
+        stackView.axis = .vertical
+        stackView.alignment = .fill
+        stackView.spacing = 8
+        return stackView
+    }()
+
+    override func didMoveToSuperview() {
+        setupViews()
+    }
+
+    func setupViews() {
+        self.addSubview(stackView)
+        stackView.snp.makeConstraints { make in
+            make.left.top.equalTo(16)
+            make.right.bottom.equalTo(-16)
+        }
+    }
+
+}
